// src/types/lookup.types.ts
export interface Category {
  id: string;
  name: string;
  description?: string;
}

export interface ItemSource {
  id: string;
  name: string;
  description?: string;
}

export interface Condition {
  id: string;
  name: string;
  description?: string;
}

export interface ProcurementStatus {
  id: string;
  name: string;
  description?: string;
}

export interface UserRole {
  id: string;
  name: string;
  description?: string;
}

export interface Location {
  id: string;
  name: string;
  description?: string;
}